requirements.txt
asgiref==3.8.1
Django==4.1.13
django-cors-headers==3.14.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.7.0
pymongo==3.11.4
pytz==2024.2
sqlparse==0.2.4
typing_extensions==4.12.2
